requirements.txt
annotated-types==0.6.0
anyio==4.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
Flask==3.0.0
gigachat==0.1.8
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
Pillow==10.1.0
protobuf==4.21.12
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
telegraph==2.2.0
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
